[Fix] width, height Int -> Double로 변경

+++ Sinzak/Scene/Detail/Model/ProductsDetail.swift
@@ -33,7 +33,7 @@ struct ProductsDetail {
     let chatCnt: Int
     let myPost: Bool
     let complete: Bool
-    let width: Int
+    let width: Double
     let vertical: Int
-    let height: Int
+    let height: Double
 }
